Segregando em fila a importação do acervo artes gráficas
