+++ SME.CDEP.Aplicacao/Servicos/Interface/IServicoImportacaoArquivoAcervoArteGrafica.cs
@@ -7,13 +7,10 @@ namespace SME.CDEP.Aplicacao.Servicos.Interface
     public interface IServicoImportacaoArquivoAcervoArteGrafica
     {
         Task<ImportacaoArquivoRetornoDTO<AcervoLinhaErroDTO<AcervoArteGraficaDTO,AcervoArteGraficaLinhaRetornoDTO>,AcervoLinhaRetornoSucessoDTO>> ImportarArquivo(IFormFile file);
-        Task PersistenciaAcervo(IEnumerable<AcervoArteGraficaLinhaDTO> acervosArteGraficaLinhas);
-        void ValidarPreenchimentoValorFormatoQtdeCaracteres(IEnumerable<AcervoArteGraficaLinhaDTO> linhas);
         Task<ImportacaoArquivoRetornoDTO<AcervoLinhaErroDTO<AcervoArteGraficaDTO,AcervoArteGraficaLinhaRetornoDTO>,AcervoLinhaRetornoSucessoDTO>> ObterImportacaoPendente();
         Task<bool> RemoverLinhaDoArquivo(long id, LinhaDTO linha);
         Task<bool> AtualizarLinhaParaSucesso(long id, LinhaDTO linha);
         Task<long> AtualizarImportacao(long id, string conteudo, ImportacaoStatus? status = null);
         Task<ImportacaoArquivoRetornoDTO<AcervoLinhaErroDTO<AcervoArteGraficaDTO, AcervoArteGraficaLinhaRetornoDTO>, AcervoLinhaRetornoSucessoDTO>> ObterImportacaoPorId(long id);
-        Task CarregarDominiosArteGrafica();
     }
 }
