+++ SME.CDEP.Aplicacao/Servicos/ServicoImportacaoArquivoAcervoArteGrafica.cs
@@ -8,6 +8,7 @@ using SME.CDEP.Dominio.Constantes;
 using SME.CDEP.Dominio.Entidades;
 using SME.CDEP.Dominio.Excecoes;
 using SME.CDEP.Dominio.Extensions;
+using SME.CDEP.Infra;
 using SME.CDEP.Infra.Dados.Repositorios.Interfaces;
 using SME.CDEP.Infra.Dominio.Enumerados;
 
@@ -15,24 +16,17 @@ namespace SME.CDEP.Aplicacao.Servicos
 {
     public class ServicoImportacaoArquivoAcervoArteGrafica : ServicoImportacaoArquivoBase, IServicoImportacaoArquivoAcervoArteGrafica 
     {
-        private readonly IServicoAcervoArteGrafica servicoAcervoArteGrafica;
-        private readonly IMapper mapper;
+        private readonly IServicoMensageria servicoMensageria;
         
         public ServicoImportacaoArquivoAcervoArteGrafica(IRepositorioImportacaoArquivo repositorioImportacaoArquivo, IServicoMaterial servicoMaterial,
             IServicoEditora servicoEditora,IServicoSerieColecao servicoSerieColecao,IServicoIdioma servicoIdioma, IServicoAssunto servicoAssunto,
             IServicoCreditoAutor servicoCreditoAutor,IServicoConservacao servicoConservacao, IServicoAcessoDocumento servicoAcessoDocumento,
-            IServicoCromia servicoCromia, IServicoSuporte servicoSuporte,IServicoFormato servicoFormato,IServicoAcervoArteGrafica servicoAcervoArteGrafica, IMapper mapper,
+            IServicoCromia servicoCromia, IServicoSuporte servicoSuporte,IServicoFormato servicoFormato,IServicoMensageria servicoMensageria, IMapper mapper,
             IRepositorioParametroSistema repositorioParametroSistema)
             : base(repositorioImportacaoArquivo, servicoMaterial, servicoEditora,servicoSerieColecao, servicoIdioma, servicoAssunto, servicoCreditoAutor,
                 servicoConservacao,servicoAcessoDocumento,servicoCromia,servicoSuporte,servicoFormato, mapper,repositorioParametroSistema)
         {
-            this.servicoAcervoArteGrafica = servicoAcervoArteGrafica ?? throw new ArgumentNullException(nameof(servicoAcervoArteGrafica));
-            this.mapper = mapper ?? throw new ArgumentNullException(nameof(mapper));
-        }
-
-        public void DefinirCreditosAutores(List<IdNomeTipoDTO> creditosAutores)
-        {
-            CreditosAutores = creditosAutores;
+            this.servicoMensageria = servicoMensageria ?? throw new ArgumentNullException(nameof(servicoMensageria));
         }
 
         public async Task<ImportacaoArquivoRetornoDTO<AcervoLinhaErroDTO<AcervoArteGraficaDTO,AcervoArteGraficaLinhaRetornoDTO>,AcervoLinhaRetornoSucessoDTO>> ObterImportacaoPendente()
@@ -53,15 +47,6 @@ namespace SME.CDEP.Aplicacao.Servicos
                 return default;
 
             return await ObterRetornoImportacaoAcervo(arquivoImportado, JsonConvert.DeserializeObject<IEnumerable<AcervoArteGraficaLinhaDTO>>(arquivoImportado.Conteudo), false);
-        }
-
-        public async Task CarregarDominiosArteGrafica()
-        {
-            await CarregarTodosOsDominios();
-            
-            await ObterCreditosAutoresPorTipo(TipoCreditoAutoria.Credito);
-            
-            await ObterSuportesPorTipo(TipoSuporte.IMAGEM);
         }
 
         public async Task<bool> RemoverLinhaDoArquivo(long id, LinhaDTO linhaDoArquivo)
@@ -93,18 +78,18 @@ namespace SME.CDEP.Aplicacao.Servicos
 
             var importacaoArquivo = ObterImportacaoArquivoParaSalvar(file.FileName, TipoAcervo.ArtesGraficas, JsonConvert.SerializeObject(acervosArtesGraficasLinhas));
             
-            await CarregarDominiosArteGrafica();
-            
             var importacaoArquivoId = await PersistirImportacao(importacaoArquivo);
            
-            ValidarPreenchimentoValorFormatoQtdeCaracteres(acervosArtesGraficasLinhas);
-            
-            await PersistenciaAcervo(acervosArtesGraficasLinhas);
-            await AtualizarImportacao(importacaoArquivoId, JsonConvert.SerializeObject(acervosArtesGraficasLinhas), acervosArtesGraficasLinhas.Any(a=> a.PossuiErros) ? ImportacaoStatus.Erros : ImportacaoStatus.Sucesso);
+            await servicoMensageria.Publicar(RotasRabbit.ExecutarImportacaoArquivoAcervoDocumental, importacaoArquivoId, Guid.NewGuid());
 
-            var arquivoImportado = await repositorioImportacaoArquivo.ObterPorId(importacaoArquivoId);
-
-            return await ObterRetornoImportacaoAcervo(arquivoImportado, acervosArtesGraficasLinhas);
+            return new ImportacaoArquivoRetornoDTO<AcervoLinhaErroDTO<AcervoArteGraficaDTO,AcervoArteGraficaLinhaRetornoDTO>,AcervoLinhaRetornoSucessoDTO>()
+            {
+                Id = importacaoArquivoId,
+                Nome = importacaoArquivo.Nome,
+                TipoAcervo = importacaoArquivo.TipoAcervo,
+                DataImportacao = DateTimeExtension.HorarioBrasilia(),
+                Status = ImportacaoStatus.Pendente
+            };
         }
 
         private async Task<ImportacaoArquivoRetornoDTO<AcervoLinhaErroDTO<AcervoArteGraficaDTO,AcervoArteGraficaLinhaRetornoDTO>,AcervoLinhaRetornoSucessoDTO>> ObterRetornoImportacaoAcervo(ImportacaoArquivo arquivoImportado, IEnumerable<AcervoArteGraficaLinhaDTO> acervosArtesGraficasLinhas, bool estaImportandoArquivo = true)
@@ -254,116 +239,6 @@ namespace SME.CDEP.Aplicacao.Servicos
                 mensagemErro.Add(acervoArteGraficaLinhaDto.Descricao.Mensagem);
 
             return mensagemErro.ToArray();
-        }
-
-        public async Task PersistenciaAcervo(IEnumerable<AcervoArteGraficaLinhaDTO> acervosArtesGraficasLinhas)
-        {
-            foreach (var acervoArteGraficaLinha in acervosArtesGraficasLinhas.Where(w=> !w.PossuiErros))
-            {
-                try
-                {
-                    var acervoArteGrafica = new AcervoArteGraficaCadastroDTO()
-                    {
-                        Titulo = acervoArteGraficaLinha.Titulo.Conteudo,
-                        Codigo = acervoArteGraficaLinha.Codigo.Conteudo,
-                        CreditosAutoresIds = CreditosAutores
-                            .Where(f => acervoArteGraficaLinha.Credito.Conteudo.RemoverAcentuacaoFormatarMinusculo().FormatarTextoEmArray().Contains(f.Nome.RemoverAcentuacaoFormatarMinusculo()))
-                            .Select(s => s.Id).ToArray(),
-                        Localizacao = acervoArteGraficaLinha.Localizacao.Conteudo,
-                        Procedencia = acervoArteGraficaLinha.Procedencia.Conteudo,
-                        Ano = acervoArteGraficaLinha.Ano.Conteudo,
-                        CopiaDigital = ObterCopiaDigitalPorValorDoCampo(acervoArteGraficaLinha.CopiaDigital.Conteudo),
-                        PermiteUsoImagem = ObterAutorizaUsoDeImagemPorValorDoCampo(acervoArteGraficaLinha.PermiteUsoImagem.Conteudo),
-                        ConservacaoId = ObterConservacaoIdPorValorDoCampo(acervoArteGraficaLinha.EstadoConservacao.Conteudo),
-                        CromiaId = ObterCromiaIdPorValorDoCampo(acervoArteGraficaLinha.Cromia.Conteudo),
-                        Largura = acervoArteGraficaLinha.Largura.Conteudo,
-                        Altura = acervoArteGraficaLinha.Altura.Conteudo,
-                        Diametro = acervoArteGraficaLinha.Diametro.Conteudo,
-                        Tecnica = acervoArteGraficaLinha.Tecnica.Conteudo,
-                        SuporteId = ObterSuporteImagemIdPorValorDoCampo(acervoArteGraficaLinha.Suporte.Conteudo),
-                        Quantidade = acervoArteGraficaLinha.Quantidade.Conteudo.ObterLongoPorValorDoCampo(),
-                        Descricao = acervoArteGraficaLinha.Descricao.Conteudo,
-                    };
-                    await servicoAcervoArteGrafica.Inserir(acervoArteGrafica);
-
-                    acervoArteGraficaLinha.DefinirLinhaComoSucesso();
-                }
-                catch (Exception ex)
-                {
-                    acervoArteGraficaLinha.DefinirLinhaComoErro(ex.Message);
-                }
-            }
-        } 
-
-        public void ValidarPreenchimentoValorFormatoQtdeCaracteres(IEnumerable<AcervoArteGraficaLinhaDTO> linhas)
-        {
-            var creditos = CreditosAutores.Where(w => w.Tipo == (int)TipoCreditoAutoria.Credito).Select(s=> mapper.Map<IdNomeDTO>(s));
-            var suportesDeImagem = Suportes.Where(w => w.Tipo == (int)TipoSuporte.IMAGEM).Select(s=> mapper.Map<IdNomeDTO>(s));
-            
-            foreach (var linha in linhas)
-            {
-                try
-                {
-                    ValidarPreenchimentoLimiteCaracteres(linha.Titulo, Constantes.TITULO);
-                    ValidarPreenchimentoLimiteCaracteres(linha.Codigo, Constantes.TOMBO);
-
-                    ValidarPreenchimentoLimiteCaracteres(linha.Credito, Constantes.CREDITO);
-                    ValidarConteudoCampoListaComDominio(linha.Credito, creditos, Constantes.CREDITO);	
-
-                    ValidarPreenchimentoLimiteCaracteres(linha.Localizacao,Constantes.LOCALIZACAO);
-                    ValidarPreenchimentoLimiteCaracteres(linha.Procedencia,Constantes.PROCEDENCIA);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.Ano,Constantes.ANO);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.CopiaDigital,Constantes.COPIA_DIGITAL);
-                    ValidarPreenchimentoLimiteCaracteres(linha.PermiteUsoImagem,Constantes.AUTORIZACAO_USO_DE_IMAGEM);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.EstadoConservacao,Constantes.ESTADO_CONSERVACAO);
-                    ValidarConteudoCampoComDominio(linha.EstadoConservacao, Conservacoes, Constantes.ESTADO_CONSERVACAO);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.Cromia,Constantes.CROMIA);
-                    ValidarConteudoCampoComDominio(linha.Cromia, Cromias, Constantes.CROMIA);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.Largura,Constantes.LARGURA);
-                    ValidarPreenchimentoLimiteCaracteres(linha.Altura,Constantes.ALTURA);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.Diametro,Constantes.DIAMETRO);
-                    ValidarPreenchimentoLimiteCaracteres(linha.Tecnica,Constantes.TECNICA);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.Suporte,Constantes.SUPORTE);
-                    ValidarConteudoCampoComDominio(linha.Suporte, suportesDeImagem, Constantes.SUPORTE);
-                    
-                    ValidarPreenchimentoLimiteCaracteres(linha.Quantidade,Constantes.QUANTIDADE);
-                    ValidarPreenchimentoLimiteCaracteres(linha.Descricao,Constantes.DESCRICAO);
-                    
-                    linha.PossuiErros = PossuiErro(linha);
-                }
-                catch (Exception e)
-                {
-                    linha.DefinirLinhaComoErro(string.Format(Constantes.OCORREU_UMA_FALHA_INESPERADA_NA_LINHA_X_MOTIVO_Y, linha.NumeroLinha, e.Message));
-                }
-            }
-        }
-
-        private bool PossuiErro(AcervoArteGraficaLinhaDTO linha)
-        {
-            return linha.Titulo.PossuiErro 
-                   || linha.Codigo.PossuiErro 
-                   || linha.Credito.PossuiErro
-                   || linha.Localizacao.PossuiErro 
-                   || linha.Procedencia.PossuiErro
-                   || linha.Ano.PossuiErro
-                   || linha.CopiaDigital.PossuiErro 
-                   || linha.PermiteUsoImagem.PossuiErro 
-                   || linha.EstadoConservacao.PossuiErro
-                   || linha.Cromia.PossuiErro 
-                   || linha.Largura.PossuiErro 
-                   || linha.Altura.PossuiErro
-                   || linha.Diametro.PossuiErro 
-                   || linha.Tecnica.PossuiErro
-                   || linha.Suporte.PossuiErro
-                   || linha.Quantidade.PossuiErro
-                   || linha.Descricao.PossuiErro;
         }
 
         private async Task<IEnumerable<AcervoArteGraficaLinhaDTO>> LerPlanilha(IFormFile file)
