+++ SME.CDEP.Infra.Servicos/Rabbit/RotasRabbit.cs
@@ -20,7 +20,8 @@ namespace SME.CDEP.Infra
         public const string NotificacaoDevolucaoEmprestimoAtrasoProlongado = "cdep.emprestimo.situacao.atraso.prolongado.aviso";
         public const string NotificacaoDevolucaoEmprestimoAtrasoProlongadoUsuario = "cdep.emprestimo.situacao.atraso.prolongado.aviso.usuario";
         
-        public const string ExecutarImportacaoArquivoAcervoBibliograficoUseCase = "cdep.importacao.arquivo.acervo.bibliografico";
-        public const string ExecutarImportacaoArquivoAcervoDocumentalUseCase = "cdep.importacao.arquivo.acervo.documental";
+        public const string ExecutarImportacaoArquivoAcervoBibliografico = "cdep.importacao.arquivo.acervo.bibliografico";
+        public const string ExecutarImportacaoArquivoAcervoDocumental = "cdep.importacao.arquivo.acervo.documental";
+        public const string ExecutarImportacaoArquivoAcervoArteGrafica = "cdep.importacao.arquivo.acervo.arte.grafica";
     }
 }
