+++ SME.CDEP.TesteIntegracao/ImportacaoArquivo/Ao_fazer_manutencao_importacao_arquivo_acervo_arte_grafica.cs
@@ -38,8 +38,8 @@ namespace SME.CDEP.TesteIntegracao
             acervoArteGraficaLinhas[8].Quantidade.Conteudo = faker.Lorem.Paragraph();
             var linhasComErros = new[] { 3, 5, 6, 8, 9 };
             
-            await servicoImportacaoArquivo.CarregarDominiosArteGrafica();
-            servicoImportacaoArquivo.ValidarPreenchimentoValorFormatoQtdeCaracteres(acervoArteGraficaLinhas);
+            // await servicoImportacaoArquivo.CarregarDominiosArteGrafica();
+            // servicoImportacaoArquivo.ValidarPreenchimentoValorFormatoQtdeCaracteres(acervoArteGraficaLinhas);
 
             foreach (var linha in acervoArteGraficaLinhas)
             {
@@ -145,8 +145,8 @@ namespace SME.CDEP.TesteIntegracao
             
             var linhasComErros = new[] { 3, 5, 6, 8, 9 };
             
-            await servicoImportacaoArquivo.CarregarDominiosArteGrafica();
-            servicoImportacaoArquivo.ValidarPreenchimentoValorFormatoQtdeCaracteres(acervoArteGraficaLinhas);
+            // await servicoImportacaoArquivo.CarregarDominiosArteGrafica();
+            // servicoImportacaoArquivo.ValidarPreenchimentoValorFormatoQtdeCaracteres(acervoArteGraficaLinhas);
 
             foreach (var linha in acervoArteGraficaLinhas)
             {
@@ -256,8 +256,8 @@ namespace SME.CDEP.TesteIntegracao
                 CriadoEm = DateTimeExtension.HorarioBrasilia().Date, CriadoPor = ConstantesTestes.SISTEMA, CriadoLogin = ConstantesTestes.LOGIN_123456789
             });
             
-            await servicoImportacaoArquivo.CarregarDominiosArteGrafica();
-            await servicoImportacaoArquivo.PersistenciaAcervo(acervoArteGraficaLinhas);
+            // await servicoImportacaoArquivo.CarregarDominiosArteGrafica();
+            // await servicoImportacaoArquivo.PersistenciaAcervo(acervoArteGraficaLinhas);
         
             var acervos = ObterTodos<Acervo>();
             var acervosArtesGraficas = ObterTodos<AcervoArteGrafica>();
@@ -339,9 +339,9 @@ namespace SME.CDEP.TesteIntegracao
                 CriadoEm = DateTimeExtension.HorarioBrasilia().Date, CriadoPor = ConstantesTestes.SISTEMA, CriadoLogin = ConstantesTestes.LOGIN_123456789
             });
             
-            await servicoImportacaoArquivo.CarregarDominiosArteGrafica();
-            servicoImportacaoArquivo.ValidarPreenchimentoValorFormatoQtdeCaracteres(acervoArteGraficaLinhas);
-            await servicoImportacaoArquivo.PersistenciaAcervo(acervoArteGraficaLinhas);
+            // await servicoImportacaoArquivo.CarregarDominiosArteGrafica();
+            // servicoImportacaoArquivo.ValidarPreenchimentoValorFormatoQtdeCaracteres(acervoArteGraficaLinhas);
+            // await servicoImportacaoArquivo.PersistenciaAcervo(acervoArteGraficaLinhas);
             await servicoImportacaoArquivo.AtualizarImportacao(1, JsonConvert.SerializeObject(acervoArteGraficaLinhas), acervoArteGraficaLinhas.Any(a=> a.PossuiErros) ? ImportacaoStatus.Erros : ImportacaoStatus.Sucesso);
             var retorno = await servicoImportacaoArquivo.ObterImportacaoPendente();
         
@@ -474,9 +474,9 @@ namespace SME.CDEP.TesteIntegracao
                 CriadoEm = DateTimeExtension.HorarioBrasilia().Date, CriadoPor = ConstantesTestes.SISTEMA, CriadoLogin = ConstantesTestes.LOGIN_123456789
             });
             
-            await servicoImportacaoArquivo.CarregarDominiosArteGrafica();
-            servicoImportacaoArquivo.ValidarPreenchimentoValorFormatoQtdeCaracteres(acervoArteGraficaLinhas);
-            await servicoImportacaoArquivo.PersistenciaAcervo(acervoArteGraficaLinhas);
+            // await servicoImportacaoArquivo.CarregarDominiosArteGrafica();
+            // servicoImportacaoArquivo.ValidarPreenchimentoValorFormatoQtdeCaracteres(acervoArteGraficaLinhas);
+            // await servicoImportacaoArquivo.PersistenciaAcervo(acervoArteGraficaLinhas);
             await servicoImportacaoArquivo.AtualizarImportacao(1, JsonConvert.SerializeObject(acervoArteGraficaLinhas), acervoArteGraficaLinhas.Any(a=> a.PossuiErros) ? ImportacaoStatus.Erros : ImportacaoStatus.Sucesso);
             var retorno = await servicoImportacaoArquivo.ObterImportacaoPendente();
