+++ SME.CDEP.Worker/WorkerRabbitMQ.cs
@@ -42,8 +42,9 @@ namespace SME.CDEP.Worker
             Comandos.Add(RotasRabbit.NotificacaoDevolucaoEmprestimoAtrasoProlongado, new ComandoRabbit("Notificação de empréstimo com atraso prolongado", typeof(INotificacaoDevolucaoEmprestimoAtrasoProlongadoUseCase), true));
             Comandos.Add(RotasRabbit.NotificacaoDevolucaoEmprestimoAtrasoProlongadoUsuario, new ComandoRabbit("Notificação usuário sobre o empréstimo em atraso prolongado", typeof(INotificacaoDevolucaoEmprestimoAtrasadoProlongadoUsuarioUseCase), true));
             
-            Comandos.Add(RotasRabbit.ExecutarImportacaoArquivoAcervoBibliograficoUseCase, new ComandoRabbit("Realiza a importação de acervos bibliográficos por planilha", typeof(IExecutarImportacaoArquivoAcervoBibliograficoUseCase), true));
-            Comandos.Add(RotasRabbit.ExecutarImportacaoArquivoAcervoDocumentalUseCase, new ComandoRabbit("Realiza a importação de acervos documental por planilha", typeof(IExecutarImportacaoArquivoAcervoDocumentalUseCase), true));
+            Comandos.Add(RotasRabbit.ExecutarImportacaoArquivoAcervoBibliografico, new ComandoRabbit("Realiza a importação de acervos bibliográficos por planilha", typeof(IExecutarImportacaoArquivoAcervoBibliograficoUseCase), true));
+            Comandos.Add(RotasRabbit.ExecutarImportacaoArquivoAcervoDocumental, new ComandoRabbit("Realiza a importação de acervos documental por planilha", typeof(IExecutarImportacaoArquivoAcervoDocumentalUseCase), true));
+            Comandos.Add(RotasRabbit.ExecutarImportacaoArquivoAcervoArteGrafica, new ComandoRabbit("Realiza a importação de acervos artes gráficas por planilha", typeof(IExecutarImportacaoArquivoAcervoArteGraficaUseCase), true));
         }
     }
 }
