+++ scripts/V48__ALTER_AUMENTAR_CAMPO_NOME_ASSUNTO_PARA_270.sql
@@ -3,3 +3,6 @@ ALTER TABLE public.assunto ALTER COLUMN nome TYPE varchar(270);
 
 --> Alterar tamanho do campo tipo autoria
 ALTER TABLE public.acervo_credito_autor ALTER COLUMN tipo_autoria TYPE varchar(30);
+
+--> Alterar tamanho do campo 'nome' de crédito/autor
+ALTER TABLE public.credito_autor ALTER COLUMN nome TYPE varchar(330);
